// src/vs/platform/agentHost/node/codex/protocol/generated/v2/ConfigWriteResponse.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.153.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/ConfigWriteResponse.ts

import type { AbsolutePathBuf } from "../AbsolutePathBuf.js";
import type { OverriddenMetadata } from "./OverriddenMetadata.js";
import type { WriteStatus } from "./WriteStatus.js";

export type ConfigWriteResponse = {
	status: WriteStatus, version: string,
	/**
	 * Canonical path to the config file that was written.
	 */
	filePath: AbsolutePathBuf, overriddenMetadata: OverriddenMetadata | null,
};